requirements.txt
discord.py>=2.0.0
PyNaCl>=1.4.0
python-dotenv>=1.0.0
ffmpeg-python>=0.2.0 